+++ app/src/androidTest/java/com/futsch1/medtimer/AndroidTestHelper.java
@@ -1,28 +1,72 @@
 package com.futsch1.medtimer;
 
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.isRoot;
+import static androidx.test.espresso.matcher.ViewMatchers.withContentDescription;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static org.hamcrest.Matchers.allOf;
 
 import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewParent;
 
 import androidx.test.espresso.UiController;
 import androidx.test.espresso.ViewAction;
+import androidx.test.espresso.ViewInteraction;
 
+import org.hamcrest.Description;
 import org.hamcrest.Matcher;
+import org.hamcrest.TypeSafeMatcher;
 
 public class AndroidTestHelper {
     public static ViewAction waitFor(long delay) {
         return new ViewAction() {
-            @Override public Matcher<View> getConstraints() {
+            @Override
+            public Matcher<View> getConstraints() {
                 return isRoot();
             }
 
-            @Override public String getDescription() {
+            @Override
+            public String getDescription() {
                 return "wait for " + delay + "milliseconds";
             }
 
-            @Override public void perform(UiController uiController, View view) {
+            @Override
+            public void perform(UiController uiController, View view) {
                 uiController.loopMainThreadForAtLeast(delay);
             }
         };
     }
+
+    public static void navigateTo(MainMenu mainMenu) {
+        String[] menuItems = {"Overview", "Medicine", "Analysis"};
+        int[] menuIds = {R.id.overviewFragment, R.id.medicinesFragment, R.id.statisticsFragment};
+        ViewInteraction bottomNavigationItemView = onView(
+                allOf(withId(menuIds[mainMenu.ordinal()]), withContentDescription(menuItems[mainMenu.ordinal()]),
+                        isDisplayed()));
+        bottomNavigationItemView.perform(click());
+    }
+
+    public static Matcher<View> childAtPosition(
+            final Matcher<View> parentMatcher, final int position) {
+
+        return new TypeSafeMatcher<>() {
+            @Override
+            public void describeTo(Description description) {
+                description.appendText("Child at position " + position + " in parent ");
+                parentMatcher.describeTo(description);
+            }
+
+            @Override
+            public boolean matchesSafely(View view) {
+                ViewParent parent = view.getParent();
+                return parent instanceof ViewGroup && parentMatcher.matches(parent)
+                        && view.equals(((ViewGroup) parent).getChildAt(position));
+            }
+        };
+    }
+
+    public enum MainMenu {OVERVIEW, MEDICINES, ANALYSIS}
 }
